util: add default resource group for virresctrl
Generally with RDT we can create more monitoring groups than
allocation groups, but this requires you to create your monitoring
group under a small set of allocation groups.

In resctrlfs default resource group is "/sys/fs/resctrl', which
contains all 'pid's of system, This is the proper place to create
an monitoring group that has no requirement with the purpose to
explictly allocate decated amount of cache or memory bandwidth.

Signed-off-by: Wang Huaqiang <huaqiang.wang@intel.com>